refactor: extract shared POS emulator engine into pos_engine.py
